# Modified TaskCreationState.
Added new data class called FillingDetails that takes TaskFormData argument. OnShowAddNewTaskClick, UiState.taskCreationState is updated to FillingDetails with the initial TaskFormData calculated from task that is clicked and one below it. So, the initial data is now handled in TasksViewModel and a taskformData is created on the show new Task click, not on the confirm click.

# Combined all state classes in UiState file and all events classes on EventsHandler.

# Moved packages like utils, states and events, AppModule, and database utils into new package 'core'

# UI
 - Modified context menu/long-press. Added icon for showing context menu.
 - Some refactoring